requirements.txt
bidict==0.22.1
compress_pickle==2.1.0
graphviz==0.20.1
guppy3==3.1.4.post1
intervaltree==3.1.0
langchain==0.2.6
langchain-openai==0.1.13
Levenshtein==0.25.1
lightning==2.2.3
marisa_trie==1.1.0
matplotlib==3.7.1
more-itertools==10.3.0
networkx==3.1
numerizer==0.2.3
numpy==1.26.4
optuna==3.6.1
pandas==2.2.2
protobuf==3.20.0
pydot==1.4.2
PyGObject==3.44.1
python-dotenv==1.0.1
rdflib==7.0.0
rpyc==6.0.0
sentencepiece==0.2.0
spacy[cuda12x]==3.7.2
spacy-dbpedia-spotlight==0.2.6
SPARQL_Burger==1.0.2
SPARQLWrapper==2.0.0
sqlitedict==2.1.0
stanza==1.7.0
tensorboard==2.17.0
torch==2.3.0
transformers==4.40.1
treelib==1.6.4
xdot==1.2
xxhash==3.4.1
z3-solver==4.12.2.0
zstandard==0.21.0